chore: rename 1a4v2 to 1a4
